requirements.txt
# Telegram Bot
aiogram>=3.13.0
python-dotenv>=1.0.0

# Excel parsing
openpyxl>=3.1.5
xlrd>=2.0.2
pandas>=2.0.0

# PostgreSQL
asyncpg>=0.29.0

# SQL Server (для обратной совместимости)
sqlalchemy>=2.0.0
aioodbc>=0.4.0
pyodbc>=5.0.0

# API для desktop приложения
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Desktop приложение
PyQt6>=6.6.0
requests>=2.31.0
